+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# Release 0.15.0
+
+### Breaking changes
+
+* For compatibility with PennyLane v0.15, the `analytic` keyword argument
+  has been removed from all devices. Analytic expectation values can
+  still be computed by setting `shots=None`.
+  [(#16)](https://github.com/XanaduAI/pennylane-orquestra/pull/16)
+
+### Contributors
+
+This release contains contributions from (in alphabetical order):
+
+Josh Izaac
+
+---
+
 # Release 0.13.1
 
 ### Bug fixes